refactor: rename `taprootutils` to `bip341`
