+++ src/payments/bip341.d.ts
